Update ru.php
The last time I offered my update for the Russian language, after a while it was reset back to the translation with errors. Please make sure that this update will not reset, I spend quite a lot of time on it. But in general, here I corrected a few grammatical errors and translated the untranslated.